telegram-bot/requirements.txt
python-telegram-bot==20.7
openai==1.42.0
python-dotenv==1.0.0
redis==5.0.1
gtts==2.4.0
SpeechRecognition==3.10.0
pydub==0.25.1
aiofiles==23.2.1
